Fucking love Mono.Cecil
It is infinitely better than those shitty Reflection APIs exposed by .NET.

I was banging my head on the wall for 2 days trying to read .winmd files,
and Mono.Cecil does it with 10 minutes of research and modification to the
script.

For Christ's sake, PLEASE use this library. It's amazing.